Implement methods related to random numbers (#6)
* Implement both overloads for Random()

* Throw NotImplementedExceptions for unimplemented Random methods

* Implement the RandomGaussian() method

The implementation uses the Box-muller transformation to accomplish
the random Gaussian implementation.

* Rename private static InternalRandom to internalRandom

* Put the internalRandom variable declaration within the Random region

* Implement noise-related methods.

The implementations are based on Perlin noise, just like the methods
from the original processing library. They still lack commentary.

* Add descriptive commentary to all Random functions

The math used in the NoiseFsc function is still not explained.
Unfortunately, the original implementation did not go over it either.